+++ ui/src/components/LineageComponent/CommonLineageComponent/index.css
